move clock types into worker
